Move URL handler registration into Core
Also declare the ckan scheme in the macOS app bundle.
