+++ Core/Properties/Resources.ja-JP.resx
@@ -208,4 +208,8 @@ https://github.com/KSP-CKAN/NetKAN/issues/new/choose
   <data name="RelationshipTypeConflicts" xml:space="preserve">
     <value>衝突</value>
   </data>
+  <data name="URLHandlersPrompt" xml:space="preserve">
+    <value>CKANはckan:// URLsハンドラを追加する権限が必要です。
+CKANにこの動作をする許可を与えますか？ いいえをクリックした場合はこのメッセージは再び表示されません。</value>
+  </data>
 </root>
