+++ Core/Properties/Resources.ko-KR.resx
@@ -202,4 +202,8 @@ Owning Mod: {2}</value>
   <data name="RelationshipTypeConflicts" xml:space="preserve">
     <value>충돌함</value>
   </data>
+  <data name="URLHandlersPrompt" xml:space="preserve">
+    <value>CKAN은 ckan:// URLs에 핸들러를 추가할 권한이 필요해요.
+CKAN에게 이것을 허용할 건가요? 아니요를 누르면 이 메세지를 다시 보지 않게 됩니다.</value>
+  </data>
 </root>
