+++ Core/Properties/Resources.resx
@@ -386,4 +386,6 @@ Free up space on that device or change your settings to use another location.
   <data name="RelationshipTypeSuggests" xml:space="preserve"><value>Suggests</value></data>
   <data name="RelationshipTypeSupports" xml:space="preserve"><value>Supports</value></data>
   <data name="RelationshipTypeConflicts" xml:space="preserve"><value>Conflicts</value></data>
+  <data name="URLHandlersPrompt" xml:space="preserve"><value>CKAN requires permission to add a handler for ckan:// URLs.
+Do you want to allow CKAN to do this? If you click no you won't see this message again.</value></data>
 </root>
