+++ GUI/Model/GUIConfiguration.cs
@@ -26,10 +26,6 @@ namespace CKAN.GUI
 
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
         [DefaultValue(false)]
-        public bool URLHandlerNoNag = false;
-
-        [JsonProperty(DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
-        [DefaultValue(false)]
         public bool CheckForUpdatesOnLaunch = false;
 
         [JsonProperty(DefaultValueHandling = DefaultValueHandling.IgnoreAndPopulate)]
